Issue: Get List of Blocks
There was a problem with creating the log.txt file. The request is to make it configurable. but with the log.txt file we also maintain the json file of the list of blocks. If we make the path configurable we have to write the configured path to a location as well. which leads to the same issue. 

Instead of that, I changed the script in such a way that, now the log file and json file is created under a directory of the home dir of the tomcat user. So we shouldn't have a permission issue now. Also the script checks now if there a folder defined already. if not, it creates the folder. before listing the blocks, it also checks if the json file is present. If not, the list blocks command also creates the json file.