Create Link to req
Works when a unit is manually made for the new package 

====
There is an issue with the Rhapsody files. My scripts are querying the rhapsody files against the name of the block to create a link. But the block is refereneced somewhere else and that file is also returned as the query result. I need to work on that query and make it more crisp.. I am udpating the Library file so the file is currently in progress
===
I fixed the issue only for the create use case. But this is a prospecting issue which may appear in other use cases as well. Especially when the model get bigger and more diagrams are added. So we must plan to extent this solution to other use cases.